add support for triggerless mentions
If the current word is greater than minCharsNoTrigger
treat it the same as @.

Clears the input buffer after a space.  If using a
trigger, then this will only happen if the input buffer
already contains a space; without a trigger, happens
regardless.  In either case, continuing to type within
settings.resetDelay will not reset the inputBuffer.