add onTriggerChar callback option
Will be called if the query is too short but begins with
triggerChar.  Passes the current string and a callback
that expects responseData in the same format as onDataRequest.
The intended use is to display an initial list of entities
for suggestion and to filter on that list even if it's
too short to trigger onDataRequest.